Src/orchestrator/requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
python-dotenv==1.2.1
httpx==0.28.1
langchain==1.2.10
langgraph==1.0.9
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-experimental==0.4.1
pandas>=2.0.0
pydantic-settings>=2.0.0